+++ test/db/lib_update_dbs_test.dart
@@ -1,7 +1,8 @@
 import 'dart:async';
+import 'dart:io';
 
 import 'package:date/date.dart';
-import 'package:elec_server/src/db/lib_iso_express.dart';
+import 'package:elec_server/db_isone.dart';
 import 'package:elec_server/src/db/lib_update_dbs.dart';
 import 'package:logging/logging.dart';
 import 'package:more/collection.dart';
@@ -9,11 +10,23 @@ import 'package:timezone/data/latest.dart';
 import 'package:dotenv/dotenv.dart' as dotenv;
 import 'package:timezone/timezone.dart';
 
-Future<void> insertDays(DailyIsoExpressReport archive, List<Date> days) async {
+Future<void> insertDays(DailyIsoExpressReport archive, List<Date> days,
+    {bool gzip = false}) async {
   await archive.dbConfig.db.open();
   for (var day in days) {
     print('Working on $day');
     await archive.downloadDay(day);
+    if (gzip) {
+      final fileName = archive.getFilename(day).path.removeSuffix('.gz');
+      if (!File(fileName).existsSync()) {
+        throw StateError('Download failed for $day');
+      }
+      var res =
+          Process.runSync('gzip', ['-f', fileName], workingDirectory: archive.dir);
+      if (res.exitCode != 0) {
+        throw StateError('GZipping the file for $day has failed!');
+      }
+    }
     await archive.insertDay(day);
   }
   await archive.dbConfig.db.close();
@@ -21,7 +34,7 @@ Future<void> insertDays(DailyIsoExpressReport archive, List<Date> days) async {
 
 Future<void> tests() async {
   // var days = Date.today(location: UTC).next.previousN(4);
-  // var days = Term.parse('1Jan24-4Jan24', UTC).days();
+  var days = Term.parse('1Jan24-29Feb24', UTC).days();
 
   ///---------------------------------------------------------------
   /// IESO
@@ -34,7 +47,7 @@ Future<void> tests() async {
   // await updateIsoneHistoricalBtmSolarArchive(Date.utc(2023, 10, 13), setUp: false);
   // await updateIsoneRtSystemLoad5minArchive(days: days, download: true);
 
-  // await insertDays(DaLmpHourlyArchive(), days);
+  await insertDays(DaLmpHourlyArchive(), days, gzip: true);
   // await updateDaEnergyOffersIsone(months: [
   //   Month.utc(2023, 1),
   //   Month.utc(2023, 2),
@@ -42,7 +55,7 @@ Future<void> tests() async {
   //   Month.utc(2023, 4),
   // ]);
 
-  await updateIsoneZonalDemand([2021], download: false);
+  // await updateIsoneZonalDemand([2021], download: false);
   // await updateIsoneZonalDemand(IntegerRange(2011, 2021));
 
   // await updateCmeEnergySettlements(days, setUp: false);